cosme: use extend
